+++ tests/test_server/test_mining_server.py
@@ -35,10 +35,10 @@ TESTS_PATH = Path(__file__).resolve().parent.parent  # path to tests directory
 def mining_client(fake_sqlalchemy_engine, model_entities, monkeypatch, tmpdir):
     """Fixture to create a client for mining_server."""
 
-    spacy_mock = Mock()
-    spacy_mock.load.return_value = model_entities
+    fake_load = Mock()
+    fake_load.return_value = model_entities
 
-    monkeypatch.setattr("bluesearch.server.mining_server.spacy", spacy_mock)
+    monkeypatch.setattr("bluesearch.server.mining_server.load_spacy_model", fake_load)
 
     # This is the original CSV file with 3 columns:
     # entity_type, model, entity_type_name
@@ -55,15 +55,10 @@ def mining_client(fake_sqlalchemy_engine, model_entities, monkeypatch, tmpdir):
     shutil.copy(models_libs, new_csv_path)
 
     # Load the model library
-    monkeypatch.setenv("BBS_DATA_AND_MODELS_DIR", str(tmpdir))
     df_csv = load_ee_models_library(tmpdir)
 
-    # The mining server expects a CSV file with the model library. So we
-    # write the data frame to a temporary CSV file
-    tmp_csv = tmpdir / "temp.csv"
-    df_csv.to_csv(tmp_csv)
     mining_server_app = MiningServer(
-        models_libs={"ee": str(tmp_csv)}, connection=fake_sqlalchemy_engine
+        models_libs={"ee": df_csv}, connection=fake_sqlalchemy_engine
     )
     mining_server_app.config["TESTING"] = True
     with mining_server_app.test_client() as client: